third_party/blink/web_tests/platform/generic/fast/forms/radio-checkbox-events-order-expected.txt

Tests the order of onclick and onchange

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Tests for checkboxes:
PASS message is ""
PASS message is "clicked"
PASS message is "changed"
PASS checkbox1.checked is true
PASS message is ""
PASS message is "clicked"
PASS message is "changed"
PASS checkbox1.checked is false
PASS message is ""
PASS message is "clicked"
PASS checkbox2.checked is false
PASS message is ""
PASS message is "clicked"
PASS message is "changed"
PASS checkbox1.checked is true
PASS message is ""
PASS message is "clicked"
PASS message is "changed"
PASS checkbox1.checked is false
PASS message is ""
PASS message is "clicked"
PASS checkbox2.checked is false
Updating |value| in click handler:
PASS message is "changed"


Tests for radio buttons:
PASS message is ""
PASS message is "clicked"
PASS message is "changed"
PASS radio1.checked is true
PASS message is ""
PASS message is "clicked"
PASS radio1.checked is true
PASS message is ""
PASS message is "clicked"
PASS radio2.checked is false
PASS message is ""
PASS message is "clicked"
PASS message is "changed"
PASS radio1.checked is true
PASS message is ""
PASS radio1.checked is true
PASS message is ""
PASS message is "clicked"
PASS radio2.checked is false
Updating |value| in click handler:
PASS message is "changed"
PASS successfullyParsed is true

TEST COMPLETE

     
